+++ CHANGELOG.md
@@ -4,12 +4,12 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 
 ## v0.0.15
 
-- **View solution for problems** ![](assets/image-2.png)
+- **View solution for problems**
 - Display a message when there is no source code
 
 ## v0.0.14
 
-- **Add support for contests** ![](assets/image.png)
+- **Add support for contests**
 - Optimize the login logic
 - Set the default file to the current active file when submitting
 - Fetch with UA