+++ CHANGELOG.md
@@ -2,9 +2,8 @@
 
 All notable changes to the "cyezoi" extension will be documented in this file.
 
-## v0.0.16
+## v0.0.17
 
-- Emergency fix for the scoreboard
-- Remove previous log
+- Fix the issue of displaying the relative time
 
 Want to see the previous log? [Click here](https://github.com/CYEZOI/cyezoi-helper/commits/main/CHANGELOG.md)